# Systems for Design and Implementation/Labs/Lab5/lab-5x-edwardinio18/server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "4.2.0",
    "title": "Cigs API",
    "description": "Cigs API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "cigsapi.onrender.com",
  "basePath": "/",
...
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/get_brand/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of the brand that we want to get",
          "type": "number"
        }
      ],
      "get": {
        "tags": ["Brands"],
        "summary": "Get a brand by a given ID",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Brand"
            }
          }
        }
      }
    },
...
    "/add_brand": {
      "post": {
        "tags": ["Brands"],
        "summary": "Add a new brand",
        "parameters": [
          {
            "name": "brand",
            "in": "body",
            "description": "Brand that we want to add",
            "schema": {
              "$ref": "#/definitions/Brand"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Brand added succesfully"
          }
        }
      }
    },
    "/add_many/{id}/cigs": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of brand that we want to add to the given cigarettes' brand's",
          "type": "number"
        },
        {
          "name": "cigs",
          "in": "body",
          "description": "List of cigarettes that we will add",
          "schema": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Cigarette"
            }
          }
        }
      ],
      "post": {
        "tags": ["Brands"],
        "summary": "Add all the given cigarettes to the brand with the given ID",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/update_brand/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of the brand that we want to update",
          "type": "number"
        },
        {
          "name": "brand",
          "in": "body",
          "description": "Brand that we want to update",
          "schema": {
            "$ref": "#/definitions/Brand"
          }
        }
      ],
      "put": {
        "tags": ["Brands"],
        "summary": "Update the brand with the given ID",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/delete_brand/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of the brand that we want to delete",
          "type": "number"
        }
      ],
      "delete": {
        "tags": ["Brands"],
        "summary": "Delete the brand with the given ID",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Cigarette": {
      "required": [
        "id",
        "name",
        "price",
        "origin",
        "photo",
        "description",
        "brand"
      ],
      "properties": {
        "name": {
          "type": "string",
          "uniqueItems": true
        },
        "price": {
          "type": "number"
        },
        "origin": {
          "type": "string"
        },
        "photo": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "brand": {
          "type": "number"
        }
      }
    },
    "Brand": {
      "required": ["name", "origin", "description", "noCountries", "photo"],
      "properties": {
        "name": {
          "type": "string",
          "uniqueItems": true
        },
        "origin": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "noCountries": {
          "type": "number"
        },
        "photo": {
          "type": "string"
        }
      }
    },
...
  }
}